Integrates upstream changes including:
- TraceStorage refactoring in main_trace.rs
- FRIE2F4 opcode and op flag
- ACE circuit module (air/src/ace.rs)
- Various processor/prover refactoring

Conflict resolutions:
- tagging module: kept deleted (our branch removed tagging)
- stack/crypto/mod.rs: kept deleted (flattened to crypto.rs; FRIE2F4 will be ported in follow-up)
- main_trace.rs: adopted TraceStorage, removed MainTraceRow struct (lives in columns.rs), kept ACE_COL_MAP accessors
- CHANGELOG.md: restored from upstream

+++ crates/lib/core/tests/crypto/sha256.rs
@@ -21,7 +21,7 @@ fn sha256_hash_bytes() {
             .map(|&bytes| u32::from_be_bytes(bytes) as u64)
             .rev()
             .collect::<Vec<u64>>(),
-        vec![length_in_bytes as u64; 1],
+        vec![length_in_bytes; 1],
     ]
     .concat();
